refactor: asset verification display
