+++ src/components/Assetlist/Asset.module.css
@@ -27,28 +27,6 @@
   font-weight: var(--font-weight-bold);
 }
 
-.author {
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  text-transform: none;
-}
-
-.author p {
-  margin-bottom: 0;
-  margin-right: calc(var(--spacer) / 2);
-}
-
-.loader {
-  display: flex;
-  align-items: center;
-  gap: calc(var(--spacer) / 2);
-}
-
-.loader span {
-  color: var(--color-blue);
-}
-
 @media screen and (min-width: 35rem) {
   .info {
     display: grid;
